#18 Enhance routing and loading components; implement path matching utility and improve product context management

+++ src/api/api.ts
@@ -1,8 +1,22 @@
 import Cookies from "js-cookie";
 import axios from "axios";
+
 const BACKEND_URL = import.meta.env.VITE_BACKEND_URL;
+
+export const isServerRunning = async () => {
+  try {
+    // It's better to use a lightweight health check endpoint if available
+    const response = await axios.get(`${BACKEND_URL}/products`);
+    return response.status === 200;
+  } catch (error: any) {
+    return false;
+  }
+};
+
 export const api = {
+  // PRODUCTS ACTIONS
   getProducts: async () => {
+    if (!(await isServerRunning())) return;
     try {
       const { data } = await axios.get(`${BACKEND_URL}/products`);
       return data;
@@ -11,6 +25,7 @@ export const api = {
     }
   },
   getProduct: async (id: string) => {
+    if (!(await isServerRunning())) return;
     try {
       const res = await axios.get(`${BACKEND_URL}/products/${id}`);
       return res;
@@ -18,7 +33,33 @@ export const api = {
       console.log("Error fetching product: " + error.message);
     }
   },
+  createProduct: async (prdct: Object) => {
+    if (!(await isServerRunning())) return;
+    try {
+      await axios.post(`${BACKEND_URL}/products`, prdct);
+    } catch (error: any) {
+      console.log("Error creating product: " + error.message);
+    }
+  },
+  deleteProduct: async (id: string) => {
+    if (!(await isServerRunning())) return;
+    try {
+      await axios.delete(`${BACKEND_URL}/products/${id}`);
+    } catch (error: any) {
+      console.log("Error deleting product: " + error.message);
+    }
+  },
+  updateProduct: async (id: string | undefined, prdc: Object) => {
+    if (!(await isServerRunning())) return;
+    try {
+      await axios.put(`${BACKEND_URL}/products/${id}`, prdc);
+    } catch (error: any) {
+      console.log("Error updating product: " + error.message);
+    }
+  },
+  // CATEGORIES ACTIONS
   getCategories: async () => {
+    if (!(await isServerRunning())) return;
     try {
       const { data } = await axios.get(`${BACKEND_URL}/categories`);
       return data;
@@ -27,6 +68,7 @@ export const api = {
     }
   },
   addCategory: async (category: object) => {
+    if (!(await isServerRunning())) return;
     try {
       await axios.post(`${BACKEND_URL}/categories`, category);
     } catch (error: any) {
@@ -34,31 +76,16 @@ export const api = {
     }
   },
   deleteCategory: async (id: string) => {
+    if (!(await isServerRunning())) return;
     try {
       await axios.delete(`${BACKEND_URL}/categories/${id}`);
     } catch (error: any) {
-      console.log("Error adding category: " + error.message);
+      console.log("Error deleting category: " + error.message);
     }
   },
-  createProduct: async (prdct: Object) => {
-    try {
-      await axios.post(`${BACKEND_URL}/products`, prdct);
-    } catch (error: any) {
-      console.log("Error creating product: " + error.message);
-    }
-  },
-  deleteProduct: async (id: string) => {
-    try {
-      await axios.delete(`${BACKEND_URL}/products/${id}`);
-    } catch (error: any) {
-      console.log("Error deleting product: " + error.message);
-    }
-  },
-  updateProduct: async (id: string | undefined, prdc: Object) => {
-    await axios.put(`${BACKEND_URL}/products/${id}`, prdc);
-  },
-  // User ACTIONS
+  // USERS ACTIONS
   registerUser: async (user: Object) => {
+    if (!(await isServerRunning())) return;
     try {
       return await axios.post(`${BACKEND_URL}/users/register`, user);
     } catch (error: any) {
@@ -66,11 +93,32 @@ export const api = {
     }
   },
   loginUser: async (user: Object) => {
-    const { data } = await axios.post(`${BACKEND_URL}/users/login`, user);
-    Cookies.set("token", data.token, { expires: 7 });
-    return data;
+    if (!(await isServerRunning())) return;
+    try {
+      const { data } = await axios.post(`${BACKEND_URL}/users/login`, user);
+      Cookies.set("token", data.token, { expires: 7 });
+      return data;
+    } catch (error: any) {
+      console.log("Error logging in user: " + error.message);
+    }
   },
+  fetchUser: async (): Promise<User | null | undefined> => {
+    if (!(await isServerRunning())) return null;
+    const token = Cookies.get("token");
+    if (!token) return null;
+    try {
+      const res = await axios.get(`${BACKEND_URL}/users/user`, {
+        headers: { Authorization: `Bearer ${token}` },
+      });
+      return res.data;
+    } catch (error: any) {
+      console.log("Error fetching user: " + error.message);
+      return null;
+    }
+  },
+  // CART ACTIONS
   addProductToCart: async (product: Product, userId: string | undefined) => {
+    if (!(await isServerRunning())) return;
     try {
       return await axios.put(`${BACKEND_URL}/users/${userId}/cart`, product);
     } catch (error: any) {
@@ -81,27 +129,11 @@ export const api = {
     productId: string,
     userId: string | undefined
   ): Promise<void> => {
+    if (!(await isServerRunning())) return;
     try {
       await axios.delete(`${BACKEND_URL}/users/${userId}/cart/${productId}`);
     } catch (error: any) {
       console.log("Error deleting product from cart: " + error.message);
-    }
-  },
-  fetchUser: async (): Promise<User | null | undefined> => {
-    const token = Cookies.get("token");
-    if (!token) return null; // Explicitly return null if no token
-    try {
-      const res = await axios.get(
-        `${import.meta.env.VITE_BACKEND_URL}/users/user`,
-        {
-          headers: { Authorization: `Bearer ${token}` },
-        }
-      );
-      return res.data; // Return the fetched user data
-    } catch (error: any) {
-      console.log(error);
-      return null;
-      // Return null in case of error
     }
   },
 };
